Refactor MenuItem to it's own file from Menu
