+++ .styleguidist/styleguidist.require.js
@@ -1,8 +1,7 @@
 var React = require('react');
 var ReactDOM = require('react-dom');
 
-var { DropdownItem } = require('../src/core/Dropdown/Dropdown');
-var { MenuItem, MenuLink } = require('../src/core/Menu/Menu');
+var { DropdownItem, MenuItem, MenuLink } = require('../src');
 global.DropdownItem = DropdownItem;
 global.MenuItem = MenuItem;
 global.MenuLink = MenuLink;
