+++ src/core/Menu/Menu.tsx
@@ -10,13 +10,14 @@ import {
   Menu as CompMenu,
   MenuProps as CompMenuProps,
   MenuListItemsProps,
-  MenuItem as CompMenuItem,
-  MenuItemProps,
-  MenuLink as CompMenuLink,
-  MenuLinkProps,
 } from '../../components/Menu/Menu';
+import {
+  MenuLanguageItem,
+  MenuLanguageItemProps,
+  MenuLanguageLink,
+  MenuLanguageLinkProps,
+} from './MenuItem';
 
-export { MenuItemProps, MenuLinkProps };
 import { Icon, IconProps } from '../Icon/Icon';
 
 type ButtonVariant = 'default' | 'language';
@@ -103,18 +104,6 @@ class MenuVariation extends Component<MenuProps> {
   }
 }
 
-export interface MenuLanguageItemProps extends MenuItemProps {
-  /** Show item as selected one */
-  selected?: boolean;
-  className?: string;
-}
-
-export interface MenuLanguageLinkProps extends MenuLinkProps {
-  /** Show item as selected one */
-  selected?: boolean;
-  className?: string;
-}
-
 /**
  * Use for dropdown menu.
  */
@@ -133,51 +122,5 @@ export class Menu extends Component<MenuProps> {
 
   render() {
     return <MenuVariation {...this.props} />;
-  }
-}
-
-export const MenuLanguageItem = ({
-  selected,
-  className,
-  ...passProps
-}: MenuLanguageItemProps) => (
-  <CompMenuItem
-    {...passProps}
-    className={classnames(className, {
-      'fi-menu-lang-item-selected': selected,
-    })}
-  />
-);
-
-export const MenuLanguageLink = ({
-  selected,
-  className,
-  ...passProps
-}: MenuLanguageLinkProps) => (
-  <CompMenuLink
-    {...passProps}
-    className={classnames(className, {
-      'fi-menu-lang-item-selected': selected,
-    })}
-  />
-);
-
-export class MenuItem extends Component<MenuItemProps> {
-  static language = (props: MenuLanguageItemProps) => {
-    return <MenuLanguageItem {...props} />;
-  };
-
-  render() {
-    return CompMenuItem;
-  }
-}
-
-export class MenuLink extends Component<MenuLinkProps> {
-  static language = (props: MenuLanguageLinkProps) => {
-    return <MenuLanguageLink {...props} />;
-  };
-
-  render() {
-    return CompMenuLink;
   }
 }
